% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_season.R
\name{date_to_season}
\alias{date_to_season}
\title{Get the two-year season belong to a particular date}
\usage{
date_to_season(x, start_month = 7, start_day = 1)
}
\arguments{
\item{x}{dates}

\item{start_month}{integer month of the season start (1=January)}

\item{start_day}{integer day of the season start}
}
\value{
integer seasons (indicating first year of two-year season)
}
\description{
Seasons are 1-year (i.e., 365 or 366) days periods beginning on some
month and day. Typically, seasons are denoted by two years. E.g.,
January 1, 2020 is in the 2019/2020 season.
}
\details{
In this codebase, a "season" is encoded by a single number, which is
the first year of the two-year season.
}
