% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_season.R
\name{ref_date_in_season}
\alias{ref_date_in_season}
\title{Get the reference date in a season}
\usage{
ref_date_in_season(
  season,
  season_start_month,
  season_start_day,
  ref_month,
  ref_day
)
}
\arguments{
\item{season}{First year of two-year season}

\item{season_start_month}{Integer month of season start (1=January)}

\item{season_start_day}{Calendar day of season start}

\item{ref_month}{Integer month of reference dates}

\item{ref_day}{Calendar day of reference dates}
}
\value{
vector of dates
}
\description{
Get the reference date in a season
}
