% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{respnet}
\alias{respnet}
\title{RESP-NET weekly hospitalization rates}
\format{
##\code{respnet}
A data frame with 1,070 rows and 3 columns:
\describe{
\item{network}{Surveillance network}
\item{week_ending}{Week ending date}
\item{rate}{Weekly hospitalization rate}
}
}
\source{
https://data.cdc.gov/Public-Health-Surveillance/Rates-of-Laboratory-Confirmed-RSV-COVID-19-and-Flu/kvib-3txy/ # nolint
}
\usage{
respnet
}
\description{
Rates of Laboratory-Confirmed RSV, COVID-19, and Flu Hospitalizations from
the RESP-NET Surveillance Systems.
These were data as of 2025-01-10.
}
\keyword{datasets}
