% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_lengths.R
\name{validate_lengths}
\alias{validate_lengths}
\title{Validate argument lengths}
\usage{
validate_lengths(...)
}
\arguments{
\item{...}{named or positional arguments}
}
\value{
\code{TRUE}, or throw error
}
\description{
All arguments must be of the same length, or they call all be of
length 1 except for a single argument
}
