% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_ceiling.R
\name{week_ceiling}
\alias{week_ceiling}
\title{Round up a date to a week-ending date}
\usage{
week_ceiling(x, week_day = 6)
}
\arguments{
\item{x}{vector of dates}

\item{week_day}{integer day of the week to round up to (1=Monday)}
}
\value{
vector of dates, of same length as \code{x}
}
\description{
Note that \code{lubridate::ceiling_date()} will round up date-time
objects, even with \code{change_on_boundary = FALSE}. This function
checks that the input is a date, not a date-time.
}
